Rename core.config.utils to core.config.endpoint
