+++ core/chain/cmd.py
@@ -19,10 +19,10 @@
 
 from typing import Optional
 
+from core.config.endpoint import get_chain_ports_from_config
 from core.config.schain.file_manager import ConfigFileManager
 from core.config.schain.main import get_skaled_container_config_path
 from core.config.schain.static_params import get_static_schain_cmd
-from core.config.utils import get_chain_ports_from_config
 from core.schains.ssl import get_ssl_filepath
 from tools.configs import SGX_SERVER_URL
 from tools.configs.containers import (
