+++ core/checks/base.py
@@ -24,16 +24,6 @@ from typing import Dict, List, Optional
 
 import statsd
 
-from core.config.schain.directory import get_schain_check_filepath
-
-from core.config.schain.file_manager import ConfigFileManager
-from core.config.utils import get_local_chain_http_endpoint_from_config
-from core.config.schain.main import (
-    get_skaled_config_rotations_ids,
-    get_upstream_config_rotation_ids,
-)
-from core.redis.chain_record import ChainRecord
-from core.firewall import IRuleController
 from core.chain.rpc import (
     check_endpoint_alive,
     check_endpoint_blocks,
@@ -42,6 +32,15 @@ from core.chain.rpc import (
 from core.chain.runner import get_container_name
 from core.chain.skaled_exit_codes import SkaledExitCodes
 from core.chain.volume import is_volume_exists
+from core.config.endpoint import get_local_chain_http_endpoint_from_config
+from core.config.schain.directory import get_schain_check_filepath
+from core.config.schain.file_manager import ConfigFileManager
+from core.config.schain.main import (
+    get_skaled_config_rotations_ids,
+    get_upstream_config_rotation_ids,
+)
+from core.firewall import IRuleController
+from core.redis.chain_record import ChainRecord
 from core.types.chain import ChainName
 from tools.configs.containers import SKALED_CONTAINER
 from tools.docker_utils import DockerUtils
@@ -49,7 +48,6 @@ from tools.helper import no_hyphens, write_json
 from tools.resources import get_statsd_client
 from tools.str_formatters import arguments_list_string
 from web.models.schain import SChainRecord
-
 
 logger = logging.getLogger(__name__)
