+++ core/monitor/schain/action_skaled.py
@@ -32,6 +32,7 @@ from core.chain.runner import (
 from core.chain.status import NodeCliStatus
 from core.chain.volume import init_data_volume
 from core.checks.schain import SkaledChecks
+from core.config.endpoint import get_base_port_from_config
 from core.config.schain.helper import (
     get_node_ips_from_config,
     get_own_ip_from_config,
@@ -40,7 +41,6 @@ from core.config.schain.main import (
     get_finish_ts_from_latest_upstream,
     get_finish_ts_from_skaled_config,
 )
-from core.config.utils import get_base_port_from_config
 from core.firewall import IRuleController
 from core.monitor.action_base import (
     CONTAINER_POST_RUN_DELAY,
