+++ tests/schains/config/config_test.py
@@ -3,6 +3,7 @@ import os
 import pytest
 
 from core.chain.volume import get_schain_volume_config
+from core.config.endpoint import get_base_port_from_config
 from core.config.schain.directory import schain_config_dir
 from core.config.schain.file_manager import ConfigFileManager
 from core.config.schain.helper import (
@@ -11,7 +12,6 @@ from core.config.schain.helper import (
     get_schain_env,
 )
 from core.config.schain.main import get_latest_finish_ts, get_rotation_ids_from_config
-from core.config.utils import get_base_port_from_config
 from tools.configs.containers import SHARED_SPACE_CONTAINER_PATH, SHARED_SPACE_VOLUME_NAME
