+++ web/routes/schains.py
@@ -23,12 +23,12 @@ from dataclasses import asdict
 from flask import Blueprint, g, request
 
 from core.chain.status import init_skaled_status
+from core.config.endpoint import get_base_port_from_config
 from core.config.schain.file_manager import ConfigFileManager
 from core.config.schain.helper import (
     get_node_ips_from_config,
     get_own_ip_from_config,
 )
-from core.config.utils import get_base_port_from_config
 from core.firewall.utils import get_default_rule_controller, get_sync_agent_ranges
 from core.schains.cleaner import get_schains_on_node
 from core.schains.ima import get_ima_version_after_migration
